// src/types/Club.ts
type Club ={ 
    id: number; 
    name: string; 
    locality: string; 
    email: string; 
    address: string; 
    training_day?: string; 
    training_time?: string; 
    description?: string; 
    contact_person?: string; 
    phone_number?: string; 
    mobile_number?: string; 
    province?: string; 
    lat?: number; 
    long?: number; }

export default Club